Did assignment 16

+++ 16_subseq/maxSeq.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-size_t maxSeq(int *array, size_t n)
+int maxSeq(int *array, size_t n)
 {
  size_t max=1;
   if(n==0)
